Refactor runtime executor for better extensibility
Extracts runtime execution logic into dedicated modules for better organization and extensibility:
- Splits ActionExecutor into base and implementation classes
- Adds support for custom executor implementations via config
- Maintains backward compatibility with existing execution logic
- Allows specifying executor class from arbitrary python modules via new core.runtime_executor config.

This change enables easier customization of runtime behavior and cleaner separation of concerns.
